requirements.txt
backcall==0.2.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
decorator==4.4.2
facenet-pytorch==2.5.1
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.26.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
httplib2==0.19.0
idna==2.10
ipykernel==5.5.0
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.11
jupyter-core==4.7.1
numpy==1.20.1
oauthlib==3.1.0
packaging==20.9
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.16
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.8.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.1
six==1.15.0
torch==1.7.1
torchvision==0.2.2.post3
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
wcwidth==0.2.5
